test: useCardNumberTest 리팩토링 - 파일 분리
